Disable AVX/SSE on unwind-c.c and unwind-dw2.c
Since %ecx can't be used for both DRAP register and __builtin_eh_return,
we must avoid stack realignment when __builtin_eh_return is used.  We
disable AVX, when the incoming stack is 16-byte aligned, or SSE, when the
incoming stack is 4-byte aligned, on unwind-c.c and unwind-dw2.c to avoid
DRAP register in 32-bit x86 libgcc.

	PR target/70439
	* config.host (tmake_file): Add i386/32/t-eh-return-no-avx for
	32-bit x86 Linux.  Add i386/32/t-eh-return-no-sse for 32-bit
	x86 Cygwin and Solaris.
	* libgcc/config/i386/32/t-eh-return-no-avx: New file.
	* libgcc/config/i386/32/t-eh-return-no-sse: Likewise.